move suspect stuff into components/issues/suspect
