+++ static/app/components/issues/suspect/consts.tsx
