+++ static/app/components/issues/suspect/useLegacyEventSuspectFlags.tsx
